Refactor out to more separate files
