+++ perceiver_pytorch/multi_perceiver_pytorch.py
@@ -1,49 +1,9 @@
 from perceiver_pytorch.perceiver_pytorch import fourier_encode
 from perceiver_pytorch.perceiver_io import PerceiverIO
+from perceiver_pytorch.modalities import InputModality, modality_encoding
 import torch
 from typing import List, Iterable, Dict, Optional, Any
 from einops import rearrange, repeat
-from dataclasses import dataclass
-
-
-@dataclass
-class InputModality:
-    name: str
-    input_channels: int
-    input_axis: int
-    num_freq_bands: int
-    max_freq: float
-    freq_base: int = 2
-    sin_only: bool = False
-    fourier_encode: bool = True
-
-    @property
-    def input_dim(self) -> int:
-        # Calculate the dimension of this modality.
-        fourier_channels = self.input_axis * ((self.num_freq_bands * 2) + 1)
-        fourier_channels = fourier_channels // 2 if self.sin_only else fourier_channels
-        input_dim = fourier_channels + self.input_channels
-        return input_dim
-
-
-def modality_encoding(batch_size: int, axes, modality_index: int, num_modalities: int) -> torch.Tensor:
-    """
-    Return one-hot encoding of modality given num_modalities, batch size and axes.
-    The result need to be compatible with the modality data for concatenation.
-    :param modality_index:
-    :param num_modalities:
-    :return:
-    """
-    one_hot = torch.eye(num_modalities, num_modalities)[modality_index]
-    to_expand = [batch_size]
-    one_hot = one_hot.unsqueeze(0)
-    for i, axis in enumerate(axes):
-        one_hot = one_hot.unsqueeze(0)
-        to_expand.append(axis)
-    to_expand.append(num_modalities)
-
-    one_hot = one_hot.expand(to_expand)
-    return one_hot
 
 
 class MultiPerceiver(torch.nn.Module):
