+++ tests/test_model.py
@@ -1,0 +1,1 @@
+from perceiver_pytorch.perceiver_io import PerceiverIO
